# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Local personal memory server: episodic (OpenSearch) + knowledge (Neo4j) derived from canonical hot JSON, archived to S3.",
        "title": "memory-mcp v2 API",
        "contact": {},
        "version": "2.0"
    },
    "host": "127.0.0.1:8420",
    "basePath": "/",
    "paths": {
...
        "/v1/{ws}/{team}/{proj}/knowledge/edges": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge"
                ],
                "summary": "Create a knowledge edge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "edge",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/knowledge.Edge"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/knowledge.EdgeResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "endpoint node not found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/{ws}/{team}/{proj}/knowledge/graph": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge"
                ],
                "summary": "Neighborhood traversal around an entity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "entity name or alias",
                        "name": "entity",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "hop depth (default 1, max 10)",
                        "name": "depth",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/knowledge.Graph"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "503": {
                        "description": "graph unavailable (§5)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/{ws}/{team}/{proj}/knowledge/nodes": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge"
                ],
                "summary": "Create a knowledge node (optionally superseding others)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "node",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/knowledge.CreateNodeInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/knowledge.NodeResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "superseded node not found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/{ws}/{team}/{proj}/knowledge/nodes/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge"
                ],
                "summary": "Purge a node from hot (S3 versioning is the backstop)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "node ULID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "must be true (§3)",
                        "name": "confirm",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/knowledge.PurgeResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "confirm missing",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "409": {
                        "description": "node is still active — archive or deprecate it first (§3)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge"
                ],
                "summary": "Transition node state (set_state / deprecate)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "node ULID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "operation",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/knowledge.PatchNodeInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/knowledge.NodeResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "409": {
                        "description": "illegal state transition",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        },
        "/v1/{ws}/{team}/{proj}/knowledge/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "knowledge"
                ],
                "summary": "Full-text search over knowledge nodes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "workspace",
                        "name": "ws",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "team",
                        "name": "team",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "project",
                        "name": "proj",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "query text",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "boolean",
                        "description": "opt-in archived/deprecated (§7)",
                        "name": "include_archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/httpserver.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/knowledge.Node"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    },
                    "503": {
                        "description": "graph unavailable (§5)",
                        "schema": {
                            "$ref": "#/definitions/httpserver.Envelope"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apierr.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "httpserver.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/apierr.Error"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "knowledge.CreateNodeInput": {
            "type": "object",
            "properties": {
                "aliases": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "body": {
                    "type": "string"
                },
                "kind": {
                    "$ref": "#/definitions/knowledge.NodeKind"
                },
                "name": {
                    "type": "string"
                },
                "provenance": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "review_after": {
                    "type": "string"
                },
                "supersedes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "trust": {
                    "$ref": "#/definitions/knowledge.Trust"
                }
            }
        },
        "knowledge.Edge": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "from": {
                    "type": "string"
                },
                "provenance": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rel": {
                    "$ref": "#/definitions/knowledge.Rel"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "knowledge.EdgeResult": {
            "type": "object",
            "properties": {
                "degraded": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "edge": {
                    "$ref": "#/definitions/knowledge.Edge"
                }
            }
        },
        "knowledge.Graph": {
            "type": "object",
            "properties": {
                "edges": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/knowledge.Edge"
                    }
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/knowledge.Node"
                    }
                }
            }
        },
...
        "knowledge.NodeKind": {
            "type": "string",
            "enum": [
                "entity",
                "fact",
                "lesson",
                "preference",
                "document"
            ],
            "x-enum-varnames": [
                "KindEntity",
                "KindFact",
                "KindLesson",
                "KindPreference",
                "KindDocument"
            ]
        },
        "knowledge.NodeResult": {
            "type": "object",
            "properties": {
                "degraded": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "node": {
                    "$ref": "#/definitions/knowledge.Node"
                }
            }
        },
        "knowledge.PatchNodeInput": {
            "type": "object",
            "properties": {
                "op": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "state": {
                    "$ref": "#/definitions/knowledge.State"
                }
            }
        },
        "knowledge.PurgeResult": {
            "type": "object",
            "properties": {
                "degraded": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "purged_id": {
                    "type": "string"
                },
                "removed_edges": {
                    "type": "integer"
                }
            }
        },
        "knowledge.Rel": {
            "type": "string",
            "enum": [
                "relates_to",
                "derived_from",
                "supersedes",
                "about"
            ],
            "x-enum-varnames": [
                "RelRelatesTo",
                "RelDerivedFrom",
                "RelSupersedes",
                "RelAbout"
            ]
        },
        "knowledge.State": {
            "type": "string",
            "enum": [
                "active",
                "archived",
                "deprecated"
            ],
            "x-enum-varnames": [
                "StateActive",
                "StateArchived",
                "StateDeprecated"
            ]
        },
        "knowledge.Trust": {
            "type": "string",
            "enum": [
                "user-stated",
                "agent-inferred",
                "imported"
            ],
            "x-enum-varnames": [
                "TrustUserStated",
                "TrustAgentInferred",
                "TrustImported"
            ]
        },
...
    }
}